src/requirements/base.txt
Django==1.9.4
django-hosts==1.2
psycopg2==2.6.1
Pillow==3.0.0
gunicorn==19.4.5
django-debug-toolbar
django-extensions
django-widget-tweaks
django-webpack-loader
sparkpost==1.2.0
django-admin-sortable
djangorestframework==3.4.0